+++ src/python/DAS/services/dbs/dbs_service.py
@@ -245,6 +245,7 @@ class DBSService(DASAbstractService):
                     pass
             if  value:
                 kwds['query'] = "find dataset, datatype, dataset.status, \
+dataset.createdate, dataset.createby, dataset.moddate, dataset.modby, \
 sum(block.numfiles), sum(block.numevents), count(block), sum(block.size) \
 where %s" % value[4:]
             else: